Supplemented drag & drop support. Support is limited due to a bug in MAUI, which on Windows seems to originate from a EinUI bug.
Details on MAUI issues with drag & drop:
Due to a bug in MAUI, the application crashes on Windows whenever something is dragged from another application and dropped to a control. When drag & drop is performed within the application, this works. Due to this bug, some essential drop functionality is deactivated (especially for dragging a file or file patth into the file entry), as it only make sense to drag a file from another application. Drop into verification entry is for now enabled because user can for example drag content from various hash values displayed in the application (I also hope that the bug will be resolved in future releases of MAUI). More information about the bug can be found here:
https://github.com/dotnet/maui/issues/6080
https://github.com/dotnet/maui/issues/6684